✨ add cancel button, upload logs, split up code
