+++ tsconfig.json
@@ -10,6 +10,7 @@
     "skipLibCheck": true,
     "moduleResolution": "node",
     "outDir": "./dist",
+    "sourceMap": true
   },
   "include": [
     "./src/**/*.ts"
